refactor: recreate asset-transfer-basic as single project
Multi-project architectures are too challenging to build properly with
Fabric.
