--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "hypernate"]
-	path = asset-transfer-basic/hypernate
-	url = git@github.com:ftsrg/hypernate.git
