+++ asset-transfer-basic/.gitattributes
@@ -1,3 +1,6 @@
+#
+# https://help.github.com/articles/dealing-with-line-endings/
+#
 # Linux start script should use lf
 /gradlew        text eol=lf
 
@@ -6,3 +9,4 @@
 
 # Binary files should be left untouched
 *.jar           binary
+
