+++ asset-transfer-basic/.idea/google-java-format.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="GoogleJavaFormatSettings">
-    <option name="enabled" value="false" />
+    <option name="enabled" value="true" />
   </component>
 </project>
