--- asset-transfer-basic/buildSrc/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-    `kotlin-dsl`
-    alias(libs.plugins.spotless)
-}
-
-repositories {
-    gradlePluginPortal()
-}
