--- asset-transfer-basic/buildSrc/settings.gradle.kts
@@ -1,7 +1,0 @@
-dependencyResolutionManagement {
-    versionCatalogs {
-        create("libs") { from(files("../gradle/libs.versions.toml")) }
-    }
-}
-
-rootProject.name = "buildSrc"
