--- asset-transfer-basic/buildSrc/src/main/kotlin/buildlogic.java-application-conventions.gradle.kts
@@ -1,4 +1,0 @@
-plugins {
-    id("buildlogic.java-common-conventions")
-    application
-}
