--- asset-transfer-basic/buildSrc/src/main/kotlin/buildlogic.java-library-conventions.gradle.kts
@@ -1,4 +1,0 @@
-plugins {
-    id("buildlogic.java-common-conventions")
-    `java-library`
-}
