+++ asset-transfer-basic/settings.gradle.kts
@@ -1,6 +1,1 @@
-plugins {
-    id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
-}
-
 rootProject.name = System.getenv("CHAINCODE_NAME") ?: "basic"
-include("app", "hypernate:lib")
